+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 3.9.0
 =========
 
-* Adds multisearch to the omnibox by using the keyword 'ms ' after 'cs '
+* Adds multisearch to the omnibox by using the keyword '!' after 'cs '
+* Open Options page from omnibox by using the keyword '.' after cs
 * It is now possible to add search engines from mycroftproject.com by right-clicking on the main (first) link of a search engine
 * A page action now allows users to add an open search engine from a web page when available
+* Improved the text definition of a multisearch onn the Options page
 
 3.8.9
 ---------